+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 1.37.0 (Unreleased)
+
+FEATURES:
+
+* **New Data Source**: `tencentcloud_cdn_domains` 
+
 ## 1.36.1 (June 12, 2020)
 
 ENHANCEMENTS: